+++ client/blocks/login/style.scss
@@ -51,10 +51,6 @@ $breakpoint-mobile: 782px; //Mobile size.
 	.masterbar {
 		display: none;
 	}
-
-	.wp-login__footer-links {
-		display: none;
-	}
 }
 
 .layout.is-jetpack-woo-dna-flow,
